+++ src/patterns/resolveSources.ts
@@ -2,7 +2,7 @@ import type { Dirent } from "node:fs"
 
 import type { PatternCompileOptions } from "./patternCompile.js"
 import type { Resource } from "./resource.js"
-import type { Rule } from "./rule.js"
+import type { GlobRule, Rule } from "./rule.js"
 import type { Source } from "./source.js"
 
 import { dirname, join } from "../unixify.js"
@@ -18,7 +18,7 @@ import { patternListCompile } from "./patternList.js"
  *
  * @since 0.6.0
  */
-export function ruleCompile(rule: Rule, options?: PatternCompileOptions): Rule {
+export function ruleCompile(rule: GlobRule, options?: PatternCompileOptions): Rule {
 	rule.compiled = patternListCompile({ ...options, list: rule.list })
 	return rule
 }
@@ -176,12 +176,12 @@ export function resolveSources(
 						}
 						results[ri] = {
 							error: err,
-							source: { inverted: false, path: epath, rules: [] },
+							source: { dir: relDirs[pi], inverted: false, path: epath, rules: [] },
 						}
 						check()
 						return
 					}
-					const source: Source = { inverted: false, path: epath, rules: [] }
+					const source: Source = { dir: relDirs[pi], inverted: false, path: epath, rules: [] }
 					let act: void | null | Error
 					try {
 						act = extract(source, buff!)