+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ## [Unreleased]
 
 - Rename `Rule.pattern`, `PatternCache.patternContext` to `list`.
+- Add `CustomRule`. Split `Rule` into `GlobRule` and `CustomRule`.
 - Fix `match.ignored`. Were always false.
 - `invert` now supports the value `2`, which enables all paths.
 - `Source.rules` is now an array of rules.